requirements.txt
django-debug-toolbar
django-filter
django>=4.0
djangorestframework
pytest
pytest-mock
PyYAML
celery[redis]
redis
drf-spectacular
djangorestframework-api-key
